GH-776 fixed tests
